backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1
python-docx==1.1.0
pdfplumber==0.10.3
PyYAML==6.0.1
chardet==5.2.0
redis==5.0.1
langdetect==1.0.9
rich==13.7.0
yt-dlp>=2024.1.0